requirements.txt
typing
tokenizers==0.10.3
nltk==3.5
tqdm==4.57.0
rouge==1.0.0
psutil==5.8.0
scikit-learn==0.24.2
prettytable==2.0.0
transformers==4.8.2
numpy==1.20.2
dataclasses
accelerate==0.3.0
tree_sitter==0.19.0
antlr4-python3-runtime==4.9.2
antlr4-tools
pytorch-lightning==1.6.5
pytorch-lightning-spells==0.1.0
wandb
